Code refactoring and performance improvements
* added data mapping to seperate classes
* introduced number connection count to player table
* cleanup of code
* added player controller

+++ TF47-API/Database/Models/GameServer/Notes.cs
@@ -14,7 +14,7 @@ namespace TF47_API.Database.Models.GameServer
     public class Note
     {
         [Key]
-        public uint NoteId { get; set; }
+        public long NoteId { get; set; }
 
         [MaxLength(2000)]
         public string Text { get; set; }
